BIG FIXES!
Fix the mixup of incorrect linking with api's and unnessary comments clogging up the code base time!

+++ src/numberlettercountfetching/IntFetchRequest.java
@@ -1,20 +1,16 @@
 package numberlettercountfetching;
-import java.util.Arrays;
 
+import java.util.Arrays;
 
 public class IntFetchRequest extends FetchRequest {
 	private int data;
 
 	public IntFetchRequest(int number) {
-		super(Arrays.asList(number)); // Pass as List
+		super(Arrays.asList(number));
 		this.data = number;
 	}
 
 	public int getIntData() {
-		return (Integer) data;
+		return data;
 	}
-
-	//    public FetchRequest(int data) {
-	//    	return this.data;
-	//    }
 }
